Updated Collisions & Improved Map Loading (see description)
Tried to update module collisions to assign map colliders from xml file but didn't worked. I'll let a function there that tries to do that, I have to fix it

Added module collisions into app so it's loaded now

The function Load() at the Map.cpp worked with a const char* (Thanks Rick for the present) that was used during all function to load a map. But there was a p2SString (wasn't well written, I had to erase one of the parameters to make it work, you know, Rick's humor...) that could be used to load that file_name by using a GetString() method instead of using all time the const char*... :)

And i'm tired of trying to make the colliders work, how the hell can I read colliders from the xml??? Now i'm complaining but I don't know why I have the feeling that making them not traversable will be more difficult :(