Updated map, render and player modules (see description, this one is important)
Updated player: now it starts at the position of camera.x/2 & camera.y/2. I have also commented some lines about precollision and Yvel because it's not finished and i didn't wanted that to disturb

Updated Render: Erased a useful define of vsync (thanks rick for that gift) and make the camera to follow the player's movement. To do that, I just update camera.pos at postupdate and make it to follow player's movement.

Updated map: Now it loads all the map at memory but it only blits what is it inside the camera. This is working but it needs to improve:
    first, we need to implement the tile
    flipping to make all tiles blit and not
    let some tiles in black

    second, we need to improve the
    player and camera initial position
    (easy jajaj) //Rick pls don't kill me
    because of this comment it's 2am