Updated Colliders, Player, Render & Scene
Now colliders work better but we still have to improve them, first we're gonna need the jump, i'm working on it.

Added debuging tool of loading at beginning of level 1. Also, when the game starts, it makes a save so if the player wants to load without saving before, it takes him to the beggining of level 1 (if there's a save already, it loads it)

Now we don't need to save camera pos because it follow player. Now we just save player pos.