Updated Collisions, Player & Scene (see description)
In collisions, the matrix of collisions had to be completed. Done.

In player, added jump mechanic (works) and updated collisions with walls to make them work ok, but still don't, i have to work more on it.

Changed save/load inputs from s and l keys to F5 to load and F6 to save