Added Music at the game and Added pathfinding Module :)
Music still needs to be downloaded and set ok with different tracks (main menu, lvl1...) but a track it's loaded at the scene now and picked from config document! :)
SFX not working, tried but not working. Just have to investigate how to implement them.

Pathfinding Module added. A* algorithm works great (tested in handout 8) but now the problem is with colliders and that they are loaded in a objects layer (not in a patterns one). I have mailed rick to see what solution can I put to this, but currently i'm pretty stuck (and tired, i've been working on the same all day)

Concluding: Pathfinding added and working but a problem with colliders interaction must be solved. To add the module I had to add also a SetWalkabilityMap method to Map Module, a ScreenToWorld method at Render Module (to pick the mouse pos) and also touch a little bit the Scene.