Update on making enemies pathfind. See description for details
The pathfinding works perfectly as some weeks ago, the problem are the enemies. They find the path okay but then it's hard to make them follow the path. And overall, after few pathfinds, it comes with an access violation because path is null... I don't understand, tomorrow I'll keep working on this and fix it, but we are very close!

Once done this, the next is just setting up things and maybe some logic.